Move the test to functorch
